+++ core/management/commands/_configure_helpers.py
@@ -62,6 +62,23 @@ SYSTEMS = [
             },
         ],
     },
+    {
+        "name": "Wikidata",
+        "description": "Entity lookup via Wikidata API",
+        "shareable": False,
+        "subsystems": [
+            {
+                "prefix": "RAGTIME_WIKIDATA",
+                "label": "Wikidata",
+                "fields": [
+                    ("USER_AGENT", "RAGtime/0.1 (https://github.com/rafacm/ragtime)", False),
+                    ("CACHE_TTL", "604800", False),
+                    ("DEBOUNCE_MS", "300", False),
+                    ("MIN_CHARS", "3", False),
+                ],
+            },
+        ],
+    },
 ]